Server/AIService/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
paho-mqtt
python-multipart
numpy
requests
easyocr

